pulled old code folder out of components, added chart files to old code keeping age chart in components for reference.
